+++ src/components/chromeExtensions/PopupError/styles.js
@@ -1,6 +1,6 @@
 import styled from "styled-components";
 
-import LogoImg from "../../assets/check-mark.png";
+import LogoImg from "../../../assets/check-mark.png";
 
 export const PopupPageLayout = styled.div`
   width: 430px;
