+++ src/components/features/UserSettings/index.js
@@ -2,11 +2,11 @@ import { useEffect, useState } from "react";
 import { useParams } from "react-router-dom";
 import { useDispatch, useSelector } from "react-redux";
 
-import { setModalOpen } from "../../store/slices/modalSlice";
+import { setModalOpen } from "../../../store/slices/modalSlice";
 import {
   setAccountSetting,
   getAccountSetting,
-} from "../../services/apiRequests";
+} from "../../../services/apiRequests";
 
 import * as S from "./styles";
